fastapi/requirements.txt
uvicorn
fastapi
pydantic 
sqlalchemy 
passlib
python-multipart
pydantic[email]
gunicorn
psycopg2
python-dotenv
OpenAI
